[#20241028] Decompose AccountDomainBrand
BREAKING CHANGE: AccountDomainBrand.vue has been split into
multiple components

- Extract domain header into DomainHeader component
- Extract settings toolbar into BrandSettingsBar component
- Extract browser frame into BrowserPreviewFrame component
- Extract instructions modal into InstructionsModal component
- Extract loading overlay into LoadingOverlay component
- Fix prop mutation in InstructionsModal using proper v-model pattern
- Update import paths to use @/ alias for better module resolution
- Remove unused imports to improve bundle size

This refactoring improves code maintainability, reusability, and follows
Vue.js best practices for component composition. Each extracted
component has a single responsibility and can be reused across the
application.
